+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Filtering campaigns by medium and vendor
+### Fixed
+- Dynamic find with empty types or mimes
+- Banner classification list (correct medium)
 
 ## [2.3.2] - 2023-01-31
 ### Fixed